feat_이미지 여러개 post 기능 테스트 완료

+++ src/main/java/org/example/canon/service/PostService.java
@@ -6,11 +6,13 @@ import org.example.canon.controller.request.PostFilterRequest;
 import org.example.canon.controller.request.PostRequest;
 import org.example.canon.dto.CustomOAuth2UserDTO;
 import org.example.canon.dto.PostDTO;
+import org.example.canon.entity.Image;
 import org.example.canon.entity.Post;
 import org.example.canon.entity.User;
 import org.example.canon.exception.PostDeleteDisableException;
 import org.example.canon.exception.PostEditDisableException;
 import org.example.canon.exception.PostNotFoundException;
+import org.example.canon.repository.ImagesRepository;
 import org.example.canon.repository.PostRepository;
 import org.example.canon.repository.UserRepository;
 import org.example.canon.specification.PostSpecification;
@@ -20,6 +22,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -29,8 +32,10 @@ public class PostService {
 
   private final PostRepository postRepository;
   private final UserRepository userRepository;
+  private final ImagesRepository imagesRepository;
   private final S3Uploader s3Uploader;
   private final ToolsService toolsService;
+  private final ImagesService imagesService;
 
 
   public long addPost(PostDTO postDTO, String email) {
@@ -42,6 +47,7 @@ public class PostService {
 
     //PostDTO로 넘어온 Request에 있는 Tools 를 tools 테이블에 넣기
     toolsService.saveTools(ret , postDTO.getTools());
+    imagesService.saveImages(ret, postDTO.getImages());
 
 
 
@@ -67,11 +73,20 @@ public class PostService {
                     .and(PostSpecification.filterByYear(postFilterRequest.getYear()))));
 
 
-
-    return postRepository.findAll(spec)
+    List<PostDTO>returnPosts = new ArrayList<>();
+    List<PostDTO>posts= postRepository.findAll(spec)
             .stream()
             .map(PostDTO::of)
             .toList();
+
+    for(PostDTO nPost:posts) {
+      List<Image> images = imagesService.getAllImagesByPostId(nPost.getId());
+
+      returnPosts.add(PostDTO.of(nPost, images));
+    }
+
+    return returnPosts;
+
   }
 
 
@@ -80,7 +95,16 @@ public class PostService {
 
   public List<PostDTO> getAllForUser() {
     List<Post> posts = postRepository.findAllByConfirmed();
-    return posts.stream().map(PostDTO::of).toList();
+
+    List<PostDTO>returnPosts = new ArrayList<>();
+
+    for(Post nPost:posts) {
+      List<Image> images = imagesService.getAllImagesByPostId(nPost.getId());
+
+      returnPosts.add(PostDTO.of(nPost, images));
+    }
+
+    return returnPosts;
   }
 
 
@@ -101,17 +125,21 @@ public class PostService {
   @Transactional
   public void deletePost(Long postId, CustomOAuth2UserDTO userDTO) {
     Optional<Post> post = postRepository.findById(postId);
-    String imageName = post.get().getFileName();
-    System.out.println("==="+ imageName+"===");
-    s3Uploader.deleteFile("example",imageName);
+    if (post.isPresent()) {
+      Post posts = post.get();
+      List<Image> images = imagesRepository.findAllByPost(posts);
+      for(Image image : images){ //반복문으로 하나하나 지우기
+        String imageName = image.getFileName();
+        System.out.println("===" + imageName + "===");
+        s3Uploader.deleteFile("example", imageName);
+      }
+    }
 
     if (userDTO.getEmail().equals(post.get().getUser().getEmail())) {
       postRepository.deleteById(postId);
     } else {
       throw new PostDeleteDisableException();
     }
-
-
   }
 
 
@@ -119,21 +147,21 @@ public class PostService {
   public void updatePost(Long postId, PostRequest request, CustomOAuth2UserDTO userDTO, MultipartFile image) throws IOException {
     Optional<Post> post = postRepository.findById(postId);
     if ((userDTO.getEmail().equals(post.get().getUser().getEmail())) ) {
-        if(image.getName().equals(post.get().getFileName())){
+      if(image.getName().equals(post.get().getFileName())){
 
         // 파일 뺴고 나머지 내용을 Request로 업데이트 해준다.
-          post.get().updatePostOnly(request);
+        post.get().updatePostOnly(request);
 
-        }else{
+      }else{
 
-          // 원래 파일 삭제
-          // 새로 업로드 후 URL + Request 내용으로 업데이트 해준다.
+        // 원래 파일 삭제
+        // 새로 업로드 후 URL + Request 내용으로 업데이트 해준다.
         s3Uploader.deleteFile("example", post.get().getFileName());
         String imageURL = s3Uploader.upload(image, "example");
         post.get().updatePostAndFile(request,imageURL,image.getName());
 
-        }
-      } else {
+      }
+    } else {
       throw new PostEditDisableException();
     }
